Former-commit-id: 6e8e80d7d3041cdbcd86ce61a52c15fdfd1cfdd6

+++ architecture/android.cpp
@@ -148,7 +148,7 @@ bool isRunning()
  */
 int keyOn(int pitch, int velocity)
 {
-    return gGlobal->keyOn(pitch, velocity);
+    return (int)gGlobal->keyOn(pitch, velocity);
 }
 
 /*
